+++ clang/test/CXX/temp/temp.res/temp.local/p6.cpp
@@ -163,7 +163,9 @@ struct A {
 A<0>::B a;
 }
 
-template <typename T> void shadow9() {  // expected-note{{template parameter is declared here}}
+template <typename T> void shadow() {  // expected-note{{template parameter is declared here}}
   using arr = int[1]; // expected-warning@+1 {{decomposition declarations are a C++17 extension}}
-  auto [T] = arr{}; // expected-error {{declaration of 'T' shadows template parameter}}
+  auto [
+    T // expected-error {{declaration of 'T' shadows template parameter}}
+    ] = arr{};
 }